+++ changelog.md
@@ -36,6 +36,7 @@
 * LLT-4924: Enable size reduction compiler flags
 * LLT-5078: Fix telio-proxy panic
 * LLT-5039: Make connection reporting unaffected by system time changes
+* LLT-4168: Fix telio-dns assert panics in debug mode
 
 <br>